lambda/requirements.txt
boto3
botocore
Pillow
